Create test for checking globalization rules
